Create Rate
Explanation of the updateUserRating function
TL;DR
The updateUserRating function takes two dictionaries of user ratings as arguments and returns a new dictionary with 
updated ratings. If a user has a rating in both dictionaries, the function calculates the average of the two ratings. 
If a user has a rating in only one dictionary, the function uses that rating. The function returns a dictionary with 
the updated ratings.
Explanation
The updateUserRating function takes two dictionaries of user ratings as arguments: userRatings and otherUserRatings. 
The function creates a new dictionary called updatedUserRatings to store the updated ratings.
The function then iterates through the userRatings dictionary using a for-in loop. For each user in the dictionary, 
the function checks if the user also has a rating in the otherUserRatings dictionary. If the user has a rating in both 
dictionaries, the function calculates the average of the two ratings and adds the updated rating to the 
updatedUserRatings dictionary. If the user has a rating in only one dictionary, the function uses that rating and adds 
it to the updatedUserRatings dictionary. If the user does not have a rating in either dictionary, the function does not 
add the user to the updatedUserRatings dictionary.
Finally, the function returns the updatedUserRatings dictionary.
Possible bugs

If the userRatings or otherUserRatings dictionaries have keys that are not strings, the function will throw an error.
If the userRatings or otherUserRatings dictionaries have values that are not integers, the function will throw an error.
If the userRatings or otherUserRatings dictionaries have duplicate keys, the function may not return the correct 
updated ratings.

Possible improvements

Add input validation to ensure that the userRatings and otherUserRatings dictionaries are valid.
Add an optional argument to allow the user to specify the weighting of each rating.
Add an optional argument to allow the user to specify the method for calculating the updated rating.

References

Swift Dictionaries