Create mesage
Explanation of the messageExchangeSystem function
TL;DR
The messageExchangeSystem function takes an integer timeLimit as an argument and sets a timer to end a conversation 
if both users have not agreed to continue the conversation within the time limit.
Explanation
The messageExchangeSystem function is a Swift function that takes an integer timeLimit as an argument. The function 
sets a timer to keep track of the time limit. When the timer reaches the time limit, the function checks if both users 
have agreed to continue the conversation. If both users have agreed, the function resets the timer and continues the 
conversation. If either user has not agreed, the function ends the conversation by printing "Conversation has ended".
The function uses a Timer object to keep track of the time limit. The Timer object is initialized with the time 
limit and a closure that is executed when the timer reaches the time limit. The closure checks if both users have agreed 
to continue the conversation. If both users have agreed, the timer is reset and the conversation continues. If either 
user has not agreed, the conversation ends by printing "Conversation has ended".
The function assumes that there are two users involved in the conversation and that there are variables user1Agreed 
and user2Agreed that are set to true if the respective user has agreed to continue the conversation.
Possible bugs

If the timeLimit argument is not an integer, the function will throw an error.
If the user1Agreed or user2Agreed variables are not set to true or false, the function may not work as 
expected.
If the user1Agreed and user2Agreed variables are not updated correctly, the function may not end the conversation 
when it should.

Possible improvements

Add input validation to ensure that the timeLimit argument is a valid integer.
Add an optional argument to allow the user to specify the message to be printed when the conversation ends.
Add an optional argument to allow the user to specify the number of users involved in the conversation.

References

Swift Timer class